+++ src/commands/submit.ts
@@ -3,6 +3,9 @@
 import * as vscode from 'vscode';
 import * as path from 'path';
 import { submitSolution } from '../utils/api';
+import { Languages, UserStatus } from '../shared';
+import { getSelectedLanguage } from '../utils/workspaceUtils';
+import { luoguUserManager } from '../luoguUserManager';
 
 /**
  * 提交题目答案
@@ -10,23 +13,36 @@ import { submitSolution } from '../utils/api';
 export async function submit(channel: vscode.OutputChannel, uri?: vscode.Uri): Promise<void> {
     let edtior = vscode.window.activeTextEditor;
     if (!edtior) { vscode.window.showErrorMessage('您没有打开任何文件，请重试'); return; }
+    if (luoguUserManager.getStatus() === UserStatus.SignedOut) { vscode.window.showErrorMessage('您没有登录，请重试'); return; }
     let text = edtior.document.getText();
     const filePath = edtior.document.fileName;
-    const O2: boolean = await vscode.window.showQuickPick(['是', '否'], { placeHolder: '是否开启O2优化 (非 C/C++/Pascal 切勿开启)' }).then(ans => {
+    const fileFName = path.parse(filePath).base;
+    const O2: boolean = await vscode.window.showQuickPick(['是', '否'], {
+        placeHolder: '是否开启O2优化 (非 C/C++/Pascal 切勿开启)'
+    }).then(ans => {
         if (ans === '是') {
             return true;
         } else {
             return false;
         }
     });
+    const selected = await vscode.window.showQuickPick(Languages).then((value) => {
+        const v = getSelectedLanguage(value);
+        return v === -1 ? 0 : v;
+    });
+    const id = await vscode.window.showInputBox({
+        placeHolder: '输入提交到的题目ID',
+        validateInput: s => s && s.trim() ? undefined : '输入不能为空'
+    });
+    if (!id) {
+        return;
+    }
     if (O2) {
         text = `// luogu-judger-enable-o2 \n ${text}`;
     }
-    const fileFName = path.parse(filePath).base;
-    const id = await vscode.window.showInputBox({ placeHolder: '输入提交到的题目ID' });
     try {
-        vscode.window.showInformationMessage(`${fileFName}正在提交到${id}到洛谷...`);
-        await submitSolution(id, text, 0, O2).then(rid => {
+        vscode.window.showInformationMessage(`${fileFName} 正在提交到 ${id}...`);
+        await submitSolution(id, text, selected, O2).then(rid => {
             vscode.window.showInformationMessage('提交成功');
             const url = `https://www.luogu.org/record/show?rid=${rid}`;
             let pannel = vscode.window.createWebviewPanel(`${rid}`, `${rid}`, vscode.ViewColumn.Two);
@@ -45,6 +61,7 @@ export async function submit(channel: vscode.OutputChannel, uri?: vscode.Uri): P
             pannel.webview.html = html;
         });
     } catch (error) {
+        vscode.window.showErrorMessage('提交失败');
         console.error(error);
     }
 }